feat: add tests, CI/CD, Docker dev experience, and UI components
- Backend: Add CategoryControllerTest and CategoryServiceTest (12 tests)
- Frontend: Add EmptyState.test.tsx and Badge.test.tsx (17 tests)
- Frontend: Add comprehensive UI component library (shadcn-style)
- AI: Add tests/test_api.py for FastAPI endpoints (8 tests)
- CI/CD: Add .github/workflows/ci.yml with parallel jobs
- Docker: Enhance docker-compose.yml with all 4 services
- Scripts: Add dev.sh and dev.ps1 for one-command startup
- Styling: Add globals.css with theme system and utilities

+++ backend/src/main/java/com/finsmart/domain/entity/Transaction.java
@@ -57,13 +57,13 @@ public class Transaction {
   private String normalizedMerchant;
 
   @Column(precision = 5, scale = 4)
-  private Double merchantConfidence;
+  private BigDecimal merchantConfidence;
 
   @Column(precision = 5, scale = 4)
-  private Double categoryConfidence;
+  private BigDecimal categoryConfidence;
 
   @Column(precision = 5, scale = 4)
-  private Double anomalyScore;
+  private BigDecimal anomalyScore;
 
   @Column(length = 512)
   private String notes;
